+++ CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Retry rehydrate or dehydrate device on network issues
 - Dehydrate device on second bootstrap
+- Dehydrate device when missing or untrusted
 
 ### Deprecated